% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efp_data.R
\name{convert_rowdata2df}
\alias{convert_rowdata2df}
\title{Convert eFP data to data.frame}
\usage{
convert_rowdata2df(efpdata, modeInput)
}
\arguments{
\item{efpdata}{data source}

\item{modeInput}{mode. "Absolute", "Relative", "Compare"}
}
\description{
Convert eFP data to data.frame
}
