% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasource.R
\docType{data}
\name{DATASOURCE}
\alias{DATASOURCE}
\title{DATASOURCE}
\format{A character vector with 17 variables
\describe{
  \item{Abiotic_Stress}{}
  \item{Abiotic_Stress_II}{}
  \item{Biotic_Stress}{}
  \item{Biotic_Stress_II}{}
  \item{Chemical}{}
  \item{Developmental_Map}{}
  \item{Developmental_Mutants}{}
  \item{Guard_Cell}{}
  \item{Hormone}{}
  \item{Lateral_Root_Initiation}{}
  \item{Light_Series}{}
  \item{Natural_Variation}{}
  \item{Regenaration}{}
  \item{Root}{}
  \item{Root_II}{}
  \item{Seed}{}
  \item{Tissue_Specific}{}
}}
\source{
\url{http://bbc.botany.utoronto.ca/efp/cgi-bin/efpWeb.cgi}
}
\usage{
DATASOURCE
}
\description{
DATASOURCE
}
\keyword{datasets}
