% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efp_data.R
\name{get_efp_absolute}
\alias{get_efp_absolute}
\title{Get Absolute eFP browser data}
\usage{
get_efp_absolute(primaryGene, dataSource = eFPscraper::DATASOURCE)
}
\arguments{
\item{primaryGene}{AGI code}

\item{dataSource}{dataSource}
}
\description{
Get Absolute eFP browser data
}
