% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efp_data.R
\name{get_efp_data}
\alias{get_efp_data}
\title{Get eFP browser data}
\usage{
get_efp_data(dataSource, modeInput, primaryGene, secondaryGene = "None",
  modeMask_low = "None", modeMask_stddev = "None", dry = F)
}
\arguments{
\item{dataSource}{data source}

\item{modeInput}{mode. "Absolute", "Relative", "Compare"}

\item{primaryGene}{AGI code}

\item{secondaryGene}{AGI code}

\item{modeMask_low}{modeMask_low}

\item{modeMask_stddev}{modeMask_stddev}

\item{dry}{dry run}
}
\description{
Get eFP browser data
}
