% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efp_data.R
\name{get_efp_relative}
\alias{get_efp_relative}
\title{Get Relative eFP browser data}
\usage{
get_efp_relative(dataSource = DATASOURCE[-12], primaryGene)
}
\arguments{
\item{dataSource}{dataSource}

\item{primaryGene}{AGI code}
}
\description{
Get Relative eFP browser data
}
