% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sampledata_absolute.R
\docType{data}
\name{sampledata_absolute}
\alias{sampledata_absolute}
\title{DATASOURCE}
\format{A character vector with 17 variables
\describe{
  \item{Abiotic_Stress}{}
}}
\source{
\url{http://www.hoge.jp/}
}
\usage{
sampledata_absolute
}
\description{
DATASOURCE
}
\keyword{datasets}
