% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sampledata_relative.R
\docType{data}
\name{sampledata_relative}
\alias{sampledata_relative}
\title{DATASOURCE}
\format{A character vector with 17 variables
\describe{
  \item{Abiotic_Stress}{}
}}
\source{
\url{http://www.hoge.jp/}
}
\usage{
sampledata_relative
}
\description{
DATASOURCE
}
\keyword{datasets}
