% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efp_data.R
\name{get_efp_absolute}
\alias{get_efp_absolute}
\title{Get Absolute eFP browser data}
\usage{
get_efp_absolute(dataSource = DATASOURCE[-12], primaryGene)
}
\arguments{
\item{dataSource}{dataSource}

\item{primaryGene}{AGI code}
}
\description{
Get Absolute eFP browser data
}
